+++ lib/domain/repository/user.dart
@@ -76,8 +76,8 @@ abstract class RxUser {
   /// Returns the [User.id] of this [RxUser].
   UserId get id => user.value.id;
 
-  /// Text representing the title of this [user].
-  String get title;
+  /// Returns the text representing the title of this [user].
+  String get title => contact.value?.contact.value.name.val ?? user.value.title;
 
   /// Listens to the updates of this [RxUser] while the returned [Stream] is
   /// listened to.